requirement.txt
pdfminer.six
langchain 
chromadb 
openai 
tiktoken 
sentence-transformers
